+++ Files/Memory.h
@@ -9,6 +9,6 @@ uint64 VirtToPFN(void* virt);
 void* PhysToVirt(uint64 phys);
 void* PFNToVirt(uint64 pfn);
 
-void* NewTableEntry(void* currentTable, uint64* idx, uint64 flags);
+void* NewTableEntry(void* currentTable, int* idx, uint64 flags);
 
 void FlushCaches(void* address);
